LayoutTests/media/video-interruption-with-resume-allowing-play-expected.txt


Test that play() during interruption does nothing, ending interruption allows playback to resume.

EVENT(canplaythrough)

RUN(internals.beginMediaSessionInterruption())
RUN(video.play())
100ms timer fired...
EXPECTED (video.paused == 'true') OK
RUN(internals.endMediaSessionInterruption('MayResumePlaying'))
EVENT(playing)

END OF TEST
